+++ src/main/java/br/com/caelum/carangobom/config/security/SecurityConfigurations.java
@@ -44,6 +44,8 @@ public class SecurityConfigurations extends WebSecurityConfigurerAdapter {
                 .antMatchers(HttpMethod.GET, "/users").permitAll()
                 .antMatchers(HttpMethod.GET, "/users/*").permitAll()
                 .antMatchers(HttpMethod.POST, "/users").permitAll()
+                .antMatchers(HttpMethod.PUT, "/users/*").permitAll()
+                .antMatchers(HttpMethod.DELETE, "/users/*").permitAll()
                 .antMatchers(HttpMethod.GET, "/brands").permitAll()
                 .antMatchers(HttpMethod.GET, "/brands/*").permitAll()
                 .antMatchers(HttpMethod.POST, "/auth").permitAll()
